feat(maintable): create component for emails list
